Rename Panel -> UIPanel, Pane -> Panel.
