+++ hotham/src/components/mod.rs
@@ -7,7 +7,6 @@ pub mod info;
 pub mod joint;
 pub mod material;
 pub mod mesh;
-pub mod pane;
 pub mod panel;
 pub mod parent;
 pub mod pointer;
@@ -18,6 +17,7 @@ pub mod skin;
 pub mod sound_emitter;
 pub mod transform;
 pub mod transform_matrix;
+pub mod ui_panel;
 pub mod visible;
 
 pub use animation_controller::AnimationController;
@@ -28,7 +28,6 @@ pub use info::Info;
 pub use joint::Joint;
 pub use material::Material;
 pub use mesh::Mesh;
-pub use pane::Pane;
 pub use panel::Panel;
 pub use parent::Parent;
 pub use pointer::Pointer;
@@ -39,4 +38,5 @@ pub use skin::Skin;
 pub use sound_emitter::SoundEmitter;
 pub use transform::Transform;
 pub use transform_matrix::TransformMatrix;
+pub use ui_panel::UIPanel;
 pub use visible::Visible;
